Implemented src/stirling2.{c,h} and test/xstirling2.c. The code is fully tested and works fine. It calculates Stirling numbers of the second kind.

+++ src/stirling2.h
@@ -4,7 +4,7 @@
 typedef struct Stirling2 Stirling2;
 
 void Stirling2_free(Stirling2 *self);
-Stirling2 *Stirling2_new(int nmax);
-int Stirling2_val(Stirling2 *self, int n, int k);
+Stirling2 *Stirling2_new(long unsigned nmax);
+long unsigned Stirling2_val(Stirling2 *self, long unsigned n, long unsigned k);
 
 #endif
